rename network to fork
