+++ tooling/ef_tests/blockchain/lib.rs
@@ -1,4 +1,4 @@
 mod deserialize;
-pub mod network;
+pub mod fork;
 pub mod test_runner;
 pub mod types;
